[IMP] Moved attribute set import from catalog module to base module
