+++ connector_magento/AUTHORS
@@ -21,3 +21,4 @@ Last update: Tue Jan  6 13:11:49 CET 2015
 * Matthieu Dietrich at Camptocamp
 * Florian da Costa at Akretion
 * Alberto Garcia at Factor Libre
+* Wolfgang Pichler at Callino
