+++ connector_magento/__manifest__.py
@@ -17,7 +17,7 @@
  'external_dependencies': {
      'python': ['magento'],
  },
- 'author': "Camptocamp,Akretion,Sodexis,Odoo Community Association (OCA)",
+ 'author': "Camptocamp,Akretion,Sodexis,Callino,Odoo Community Association (OCA)",
  'license': 'AGPL-3',
  'website': 'http://www.odoo-magento-connector.com',
  'images': ['images/magento_backend.png',
