+++ connector_magento/models/magento_website/common.py
@@ -67,6 +67,16 @@ class MagentoWebsite(models.Model):
         self.write({'import_partners_from_date': next_time})
         return True
 
+    @api.multi
+    def import_attributes_set(self):
+        for website in self:
+            backend = website.backend_id
+            self.env['magento.product.attributes.set'].with_delay(identity_key=identity_exact).import_batch(
+                backend,
+                filters={'magento_website_id': website.external_id}
+            )
+
+        return True
 
 class WebsiteAdapter(Component):
     _name = 'magento.website.adapter'
