+++ connector_magento_product_catalog/models/__init__.py
@@ -1,5 +1,4 @@
 from . import magento_backend
-from . import magento_website
 from . import custom_attributes
 from . import product
 from . import product_template
