--- connector_magento_product_catalog/models/magento_website/__init__.py
@@ -1,1 +1,0 @@
-from . import common
