+++ connector_magento_product_catalog/views/magento_backend_views.xml
@@ -11,15 +11,6 @@
 				position="inside">
 				<group>
 					<div>
-						<label
-							string="Import attribute sets, attributes and value"
-							class="oe_inline" />
-					</div>
-					<button name="import_attributes_set" type="object"
-						class="oe_highlight" string="Import in background" />
-				</group>
-				<group>
-					<div>
 						<label string="Export all values for binded products since"
 							class="oe_inline" />
 						<field name="export_products_from_date" class="oe_inline"
